requirements.txt
torch
torchvision
tqdm
tensorboard
Pillow
opencv-python
scipy
torch-summary
scikit-learn
scikit-image
pandas
shapely
rasterio
matplotlib
geopandas
tensorboard-logger
descartes
tensorboard-plugin-customizable-plots
seaborn